requirements.txt
opensearch-py
python-dotenv
sentence-transformers
langchain-huggingface
typer
python-slugify
python-frontmatter
gazpacho
sqlalchemy
psycopg2-binary
httpx
openai-whisper
langchain
